desktop/requirements.txt
pygetwindow>=0.0.9
pynput>=1.7.6
mss>=9.0.0
psutil>=5.9.0
pystray>=0.19.0
Pillow>=10.0.0
requests>=2.31.0
